Update the listing controlller add ->  action by user method, and update the user controller with  statistic data in dashboard template

+++ app/Http/Controllers/ListingController.php
@@ -73,6 +73,10 @@ class ListingController extends Controller
     // Update the listing
     public function update(Request $request, Listing $listing){
 
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
+
         $formData = $request->validate([
             "name" => "required|max:50",
             "title" => "required|max:50",
@@ -120,6 +124,10 @@ class ListingController extends Controller
 
     // Delete the listing
     public function destroy(Listing $listing){
+
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
         Listing::destroy($listing->id);
         return redirect()->route("dashboard")->with("message", "Listing deleted successfully");
     }
